requirements.txt
openai~=1.52.2
elevenlabs~=1.12.1
schedule~=1.2.2
requests~=2.32.3
python-dotenv~=1.0.1
podcastfy~=0.2.9

pydub~=0.25.1
PyYAML~=6.0.2
httpx~=0.27.2
feedparser~=6.0.11